requirements.txt
absl-py==0.6.1
altgraph==0.10.2
argcomplete==1.8.2
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0rc1
bdist-mpkg==0.5.0
beautifulsoup4==4.5.3
bleach==1.5.0
bonjour-py==0.3
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
Django==1.11.13
docx2txt==0.6
EbookLib==0.15
enum34==1.1.6
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
html5lib==0.9999999
idna==2.8
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.2.5
macholib==1.5.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
modulegraph==0.10.4
nose==1.3.7
numpy==1.15.3
oauthlib==2.1.0
pandas==0.23.0
pbr==5.1.0
Pillow==5.3.0
pocketsphinx==0.1.3
protobuf==3.6.1
py2app==0.7.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
PySocks==1.6.8
python-dateutil==2.7.3
python-pptx==0.6.5
python-twitter==3.5
pytz==2013.7
requests==2.21.0
requests-oauthlib==1.0.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
singledispatch==3.4.0.3
six==1.10.0
SpeechRecognition==3.6.3
subprocess32==3.5.3
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
textract==1.6.1
tornado==5.0.2
tweepy==3.7.0
urllib3==1.24.1
virtualenv==16.0.0
Werkzeug==0.14.1
xattr==0.6.4
xlrd==1.0.0
XlsxWriter==1.1.2
zope.interface==4.1.1
